==> scripts/0_first_sample.ipynb <==
# Random Sample of 100 Documents Extraction

# Load libraries and set working directory to where files are stored
import pandas as pd
import os

articles = pd.read_csv('ner_corpus_only_relevant_domains.csv')
articles_only = articles['text_clean']

training_data = []
for article in articles_only:
        training_data.append({"text": article.strip()})
len(training_data)

import random
# sample = random.sample(training_data, k=100)

import srsly
# srsly.write_jsonl("prodigy_sample.jsonl", sample) # do not overwrite the sample
sample = pd.read_json("temp/prodigy_sample.jsonl", lines = True)
